refactor(api): streamline language detection in prompt processing
- Replaced the manual language detection logic in PromptProcessorService with a call to detectLanguageFromCharacter for improved maintainability and clarity.
- Updated PilotService and SkillInvokerService to utilize the new detectLanguage function, enhancing locale preference handling.
- Ensured consistent use of optional chaining and nullish coalescing for safer property access throughout the code.
